Isolate engine to separate file; rename to wrapper
